DjangoBook/requirements.txt
Django==1.9.7
django-appconf==1.0.2
django-filter==0.13.0
django-imagekit==4.0
djangorestframework==3.4.0
Markdown==2.6.6
MySQL-python==1.2.5
olefile==0.44
pilkit==2.0
Pillow==4.1.1
six==1.10.0
South==1.0.2
